+++ packages/@expo/cli/src/start/server/metro/createServerComponentsMiddleware.ts
@@ -285,7 +285,7 @@ export function createServerComponentsMiddleware(
   }
 
   async function getExpoRouterClientReferencesAsync(
-    { platform }: { platform: string },
+    { platform, domRoot }: { platform: string; domRoot?: string },
     files: ExportAssetMap
   ): Promise<{
     reactClientReferences: string[];
@@ -295,6 +295,8 @@ export function createServerComponentsMiddleware(
     const contents = await ssrLoadModuleArtifacts(routerModule, {
       environment: 'react-server',
       platform,
+      modulesOnly: true,
+      domRoot,
     });
 
     // Extract the global CSS modules that are imported from the router.